managed to complete first disclosure over openid4vp
