+++ eudi-demo-frontend/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>GET Request Button</title>
+  <title>OpenID4VP Frontend</title>
 </head>
 <body>
-  <button id="getButton">Start OpenID4VP Session</button>
+  <button id="start-session">Start OpenID4VP Session</button>
   <script src="index.js"></script>
 </body>
 </html>
