+++ eudi/credentials/sdjwtvc/crypto.go
@@ -5,6 +5,7 @@ import (
 	"crypto/x509"
 	"encoding/json"
 	"encoding/pem"
+	"errors"
 	"os"
 
 	"github.com/golang-jwt/jwt/v5"
@@ -40,16 +41,21 @@ func (c *DefaultEcdsaJwtCreator) CreateSignedJwt(customHeaderFields map[string]s
 	return jwt, nil
 }
 
+func DecodeEcdsaPrivateKey(bytes []byte) (*ecdsa.PrivateKey, error) {
+	block, _ := pem.Decode(bytes)
+	if block == nil || block.Type != "EC PRIVATE KEY" {
+		return nil, errors.New("failed to decode ecsda private key")
+	}
+
+	return x509.ParseECPrivateKey(block.Bytes)
+
+}
+
 func ReadEcdsaPrivateKey(path string) (*ecdsa.PrivateKey, error) {
 	keyBytes, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
 
-	block, _ := pem.Decode(keyBytes)
-	if block == nil || block.Type != "EC PRIVATE KEY" {
-		return nil, err
-	}
-
-	return x509.ParseECPrivateKey(block.Bytes)
+	return DecodeEcdsaPrivateKey(keyBytes)
 }
